requirements.txt
django<3
djangorestframework
psycopg2
psycopg2-binary
autopep8
jdatetime